refactor: remove duplicate definitions
Remove types and interface definitions that
were duplicate at core and cli, choosing the
ones defined at the core.
