+++ core/mod.ts
@@ -1,1 +1,2 @@
 export * from "./types.ts";
+export * from "./plugins/mod.ts";
